// src/app/core/models/menu.class.ts
import {Role} from './role.class';

export interface Menu {
  menuId?: number;
  url?: string;
  name?: string;
  description?: string;
  menuChildren?: Menu[];
  orderPriority: number;
  createdDate?: Date;
  roleDto: Role;
  level?: number;
  expand?: boolean;
  parent?: Menu;
  parentId?: number;
  roleId?: number;
  icon?: string;
  applicationId?: number;
}